Sample: make m_nFrames and m_loop long long
Previously, both loop markers and the total number of samples were
stored in `int`. This caused the `SampleEditor` to be unusable for large
samples (with a number of frames exceeding the maximum of the platform
`int`). Now, we use `long long` and allow the sample to be as large as
possible.

Fortunately, within the XSD file and XML read/write routines, we do not
have to change anything.

+++ src/core/Basics/Instrument.cpp
@@ -855,9 +855,9 @@ void Instrument::setSample(
 	checkForMissingSamples( trigger );
 }
 
-int Instrument::getLongestSampleFrames() const
+long long Instrument::getLongestSampleFrames() const
 {
-	int nLongestFrames = 0;
+	long long nLongestFrames = 0;
 
 	for ( const auto& pComponent : *m_pComponents ) {
 		if ( pComponent != nullptr ) {
